backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic-settings==2.2.1
httpx==0.27.0
slowapi==0.1.9
cryptography==42.0.5
python-dotenv==1.0.1
